requirements.txt
lxml
pyyaml
requests
tqdm
urllib3
